Fix tmp issue?

+++ ExonCov/cli.py
@@ -123,81 +123,85 @@ class ImportBam(Command):
         db.session.add(sample)
         db.session.commit()
 
+        # Create temp_dir
+        if not temp_path:
+            temp_dir = tempfile.mkdtemp()
+        else:  # assume path exist and user is responsible for this directory
+            temp_dir = temp_path
+
         # Run sambamba
-        temp_dir = tempfile.mkdtemp(dir=temp_path)
         p = subprocess.Popen(shlex.split(sambamba_command), stdout=subprocess.PIPE)
         exon_measurement_file_path = '{0}/{1}'.format(temp_dir, sample.exon_measurement_file)
-        exon_measurement_file = open(exon_measurement_file_path, "w")
-        for line in p.stdout:
-            if print_output:
-                print line
+        with open(exon_measurement_file_path, "w") as exon_measurement_file:
+            for line in p.stdout:
+                if print_output:
+                    print line
 
-            # Header
-            if line.startswith('#'):
-                header = line.rstrip().split('\t')
-                measurement_mean_coverage_index = header.index('meanCoverage')
-                measurement_percentage10_index = header.index('percentage10')
-                measurement_percentage15_index = header.index('percentage15')
-                measurement_percentage20_index = header.index('percentage20')
-                measurement_percentage30_index = header.index('percentage30')
-                measurement_percentage50_index = header.index('percentage50')
-                measurement_percentage100_index = header.index('percentage100')
-                exon_measurement_file.write(
-                    '#{chr}\t{start}\t{end}\t{cov}\t{perc_10}\t{perc_15}\t{perc_20}\t{perc_30}\t{perc_50}\t{perc_100}\n'.format(
-                        chr='chr',
-                        start='start',
-                        end='end',
-                        cov='measurement_mean_coverage',
-                        perc_10='measurement_percentage10',
-                        perc_15='measurement_percentage15',
-                        perc_20='measurement_percentage20',
-                        perc_30='measurement_percentage30',
-                        perc_50='measurement_percentage50',
-                        perc_100='measurement_percentage100',
+                # Header
+                if line.startswith('#'):
+                    header = line.rstrip().split('\t')
+                    measurement_mean_coverage_index = header.index('meanCoverage')
+                    measurement_percentage10_index = header.index('percentage10')
+                    measurement_percentage15_index = header.index('percentage15')
+                    measurement_percentage20_index = header.index('percentage20')
+                    measurement_percentage30_index = header.index('percentage30')
+                    measurement_percentage50_index = header.index('percentage50')
+                    measurement_percentage100_index = header.index('percentage100')
+                    exon_measurement_file.write(
+                        '#{chr}\t{start}\t{end}\t{cov}\t{perc_10}\t{perc_15}\t{perc_20}\t{perc_30}\t{perc_50}\t{perc_100}\n'.format(
+                            chr='chr',
+                            start='start',
+                            end='end',
+                            cov='measurement_mean_coverage',
+                            perc_10='measurement_percentage10',
+                            perc_15='measurement_percentage15',
+                            perc_20='measurement_percentage20',
+                            perc_30='measurement_percentage30',
+                            perc_50='measurement_percentage50',
+                            perc_100='measurement_percentage100',
+                        )
                     )
-                )
 
-            # Measurement
-            else:
-                data = line.rstrip().split('\t')
-                chr, start, end = data[:3]
-                exon_id = '{0}_{1}_{2}'.format(chr, start, end)
-                measurement_mean_coverage = float(data[measurement_mean_coverage_index])
-                measurement_percentage10 = float(data[measurement_percentage10_index])
-                measurement_percentage15 = float(data[measurement_percentage15_index])
-                measurement_percentage20 = float(data[measurement_percentage20_index])
-                measurement_percentage30 = float(data[measurement_percentage30_index])
-                measurement_percentage50 = float(data[measurement_percentage50_index])
-                measurement_percentage100 = float(data[measurement_percentage100_index])
-                exon_measurement_file.write(
-                    '{chr}\t{start}\t{end}\t{cov}\t{perc_10}\t{perc_15}\t{perc_20}\t{perc_30}\t{perc_50}\t{perc_100}\n'.format(
-                        chr=chr,
-                        start=start,
-                        end=end,
-                        cov=measurement_mean_coverage,
-                        perc_10=measurement_percentage10,
-                        perc_15=measurement_percentage15,
-                        perc_20=measurement_percentage20,
-                        perc_30=measurement_percentage30,
-                        perc_50=measurement_percentage50,
-                        perc_100=measurement_percentage100,
+                # Measurement
+                else:
+                    data = line.rstrip().split('\t')
+                    chr, start, end = data[:3]
+                    exon_id = '{0}_{1}_{2}'.format(chr, start, end)
+                    measurement_mean_coverage = float(data[measurement_mean_coverage_index])
+                    measurement_percentage10 = float(data[measurement_percentage10_index])
+                    measurement_percentage15 = float(data[measurement_percentage15_index])
+                    measurement_percentage20 = float(data[measurement_percentage20_index])
+                    measurement_percentage30 = float(data[measurement_percentage30_index])
+                    measurement_percentage50 = float(data[measurement_percentage50_index])
+                    measurement_percentage100 = float(data[measurement_percentage100_index])
+                    exon_measurement_file.write(
+                        '{chr}\t{start}\t{end}\t{cov}\t{perc_10}\t{perc_15}\t{perc_20}\t{perc_30}\t{perc_50}\t{perc_100}\n'.format(
+                            chr=chr,
+                            start=start,
+                            end=end,
+                            cov=measurement_mean_coverage,
+                            perc_10=measurement_percentage10,
+                            perc_15=measurement_percentage15,
+                            perc_20=measurement_percentage20,
+                            perc_30=measurement_percentage30,
+                            perc_50=measurement_percentage50,
+                            perc_100=measurement_percentage100,
+                        )
                     )
-                )
 
-                exon_measurements[exon_id] = {
-                    'sample_id': sample.id,
-                    'exon_id': exon_id,
-                    'measurement_mean_coverage': measurement_mean_coverage,
-                    'measurement_percentage10': measurement_percentage10,
-                    'measurement_percentage15': measurement_percentage15,
-                    'measurement_percentage20': measurement_percentage20,
-                    'measurement_percentage30': measurement_percentage30,
-                    'measurement_percentage50': measurement_percentage50,
-                    'measurement_percentage100': measurement_percentage100,
-                }
+                    exon_measurements[exon_id] = {
+                        'sample_id': sample.id,
+                        'exon_id': exon_id,
+                        'measurement_mean_coverage': measurement_mean_coverage,
+                        'measurement_percentage10': measurement_percentage10,
+                        'measurement_percentage15': measurement_percentage15,
+                        'measurement_percentage20': measurement_percentage20,
+                        'measurement_percentage30': measurement_percentage30,
+                        'measurement_percentage50': measurement_percentage50,
+                        'measurement_percentage100': measurement_percentage100,
+                    }
 
         # Compress, index and rsync
-        exon_measurement_file.close()
         exon_measurement_file_path_gz = '{0}.gz'.format(exon_measurement_file_path)
         pysam.tabix_compress(exon_measurement_file_path, exon_measurement_file_path_gz)
         pysam.tabix_index(exon_measurement_file_path_gz, seq_col=0, start_col=1, end_col=2)
@@ -248,7 +252,8 @@ class ImportBam(Command):
             db.session.commit()
 
         # Remove temp_dir
-        shutil.rmtree(temp_dir)
+        if not temp_path:
+            shutil.rmtree(temp_dir)
 
 class SearchSample(Command):
     """Search sample in database."""
